+++ FslogSrc/ES.Fslog/AssemblyInfo.fs
@@ -5,10 +5,10 @@ open System.Runtime.InteropServices
 [<assembly: AssemblyTitleAttribute("Fslog - F# Logging Framework Library")>]
 [<assembly: GuidAttribute("7D77EDBF-1186-4C9C-BECB-A1F1D5305FF2")>]
 [<assembly: AssemblyProductAttribute("Fslog - F# Logging Framework")>]
-[<assembly: AssemblyVersionAttribute("1.1.0.0")>]
-[<assembly: AssemblyInformationalVersionAttribute("1.1.0.0")>]
-[<assembly: AssemblyFileVersionAttribute("1.1.0.0")>]
+[<assembly: AssemblyVersionAttribute("1.2.0.0")>]
+[<assembly: AssemblyInformationalVersionAttribute("1.2.0.0")>]
+[<assembly: AssemblyFileVersionAttribute("1.2.0.0")>]
 do ()
 
 module internal AssemblyVersionInformation =
-    let [<Literal>] Version = "1.1.0.0"
+    let [<Literal>] Version = "1.2.0.0"