+++ FslogSrc/ES.Fslog/ILogProvider.fs
@@ -4,6 +4,7 @@ open System
 
 type ILogProvider = 
     interface
+        inherit IDisposable
         abstract AddLogSourceToLoggers : LogSource -> unit
         abstract AddLogger : ILogger -> unit
     end